Renamed fuzzy matching functions.
Renamed functions to match the R equivalents, agrep and agrepl.  Note that these match on full strings unlike the R functions which do greedy matching. If the "pattern" string is shorter than the test case in the R version, then the additional length has no cost and therefore the R version overmatches when using fixed strings (the default).