Fix synonymize.py to follow sister synonyms
In TPL, the statement "there is ONE accepted name for each unique name", is
only true when one includes authors. In fact, ignoring authorship, there are
some synonyms that match to multiple accepteds (eg "Amaryllis dubia"). This
made it tricky to reverse (merge) an expanded list perfectly. The code now
handles that problem.