Changes to scripts to use myco project data
Not the cleanest way to do this, but paths and file names were hard
coded so I've changed them in this branch.  Really these scripts should
be given nice command line interfaces and then all that would need to
change for a new analysis is one bash script that called all the python
and R code. That is a todo.