Code, data for myco project thru gbif extraction
I've gone ahead and committed some of the intermediate results although
hey can be recreated by code. My feeling is that this repo is for
recording the steps in the analyses, and making a super clean repo is
less important than recording anything.  At some point, I will need to
go through the commits in this branch and cherry pick some changes for
the taxon-names-utils repository which will eventually ebcome the public
code for synonym expansion and fuzzy name matching.